Create new directory and change path
