+++ src/utils/render.js
@@ -21,15 +21,4 @@ const render = (container, element, place) => {
   }
 };
 
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours() % 12);
-  const minutes = castTimeFormat(date.getMinutes());
-
-  return `${hours}:${minutes}`;
-};
-
-export {RenderPosition, render, formatTime, createElement};
+export {RenderPosition, render, createElement};
